omgomgomg

+++ push_swap/ext_functions.c
@@ -6,13 +6,13 @@
 /*   By: kpoilly <kpoilly@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/10 17:52:34 by kpoilly           #+#    #+#             */
-/*   Updated: 2023/12/29 14:14:28 by kpoilly          ###   ########.fr       */
+/*   Updated: 2023/12/31 18:03:24 by kpoilly          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
 	int		i;
 	int		sign;
